+++ Changelog.md
@@ -2,6 +2,7 @@ WhateverGreen Changelog
 =======================
 #### v1.6.5
 - Added constants for macOS 14 support
+- Added a new boot argument `-igfxblt` to revert the optimizations done by the compiler in backlight related functions, fixing the 3-minute dark screen issue and making Backlight Smoother (BLS) work on mobile Coffee Lake platforms running macOS 13.4 or later. (by @0xFireWolf)
 
 #### v1.6.4
 - Fixed Radeon RX 5500 XT identification regression